Separate the tools routes
