+++ backend/src/routes/v0/__init__.py
@@ -1,6 +1,7 @@
 from .llm import router as llm
 from .thread import router as thread
 from .tool import router as tool
+from .tool.custom import router as tool_custom
 from .retrieve import router as retrieve
 from .source import router as source
 from .info import router as info
@@ -12,4 +13,4 @@ from .agent import router as agent
 from .model import router as model
 from .server import router as server
 
-__all__ = ["llm", "thread", "tool", "retrieve", "source", "info", "auth", "token", "storage", "settings", "agent", "model", "server"]
+__all__ = ["llm", "thread", "tool", "tool_custom", "retrieve", "source", "info", "auth", "token", "storage", "settings", "agent", "model", "server"]
